Code for final submission version
"Cosine Similarity Code [save results].ipynb" -- this file contains all of the code to generate the data. The data is all saved as Excel files, which are included in the "Data for Figures" folder.
"Cosine Similarity Code [produce figures].ipynb" -- this file loads the data and uses it to produce figures. If one downloads the data files to their desktop (or other local location), they can use this code to reproduce the figures.